refactor(stats): extract AnalysisGroupViewComponent from page shell
Der 682-Zeiler analysis-page.component.ts wird in zwei klar
abgegrenzte Komponenten zerlegt — Vorarbeit für die mat-tabs-
Aufteilung der Analyse-Seite:

- analysis-group-view.component.ts (neu) bündelt Chart, KPI-Grid,
  Heatmap, Trends und Type-Pie und liest alles aus dem injected
  AnalysisStore. Keine Inputs — wer den Selector irgendwo platziert,
  bekommt automatisch die Daten der aktiven Sicht.
- analysis-page.component.ts behält Preview-Banner, Page-Header,
  Filter-Bar, Kind-Filter und Empty-CTA. Der bisherige Inhalt
  wird durch <app-analysis-group-view /> ersetzt.
- kindDisplayName(value) zieht die $localize-Logik aus dem Page-
  Component in exercise-display-names.ts, damit Shell (Filter-Chips)
  und Group-View (Type-Pie-Legende) sich denselben Übersetzungsweg
  teilen.

Test-Setup: Mocks für Heatmap/TypePie/StatsChart/SetsDistribution
werden jetzt am AnalysisGroupViewComponent überschrieben, FilterBar
weiter an der Shell. typeBreakdownDisplay() wird über
fixture.debugElement.query(By.directive(AnalysisGroupViewComponent))
aus der Child-Instanz gelesen; alle 53 Specs bleiben grün.
