+++ web/src/app/stats/shell/analysis-page.component.ts
@@ -1,4 +1,4 @@
-import { DecimalPipe, DOCUMENT, isPlatformBrowser } from '@angular/common';
+import { DOCUMENT, isPlatformBrowser } from '@angular/common';
 import {
   Component,
   computed,
@@ -7,31 +7,21 @@ import {
   inject,
   PLATFORM_ID,
   REQUEST,
-  signal,
 } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
-import { MatButtonToggleModule } from '@angular/material/button-toggle';
 import { MatCardModule } from '@angular/material/card';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatIconModule } from '@angular/material/icon';
 import { MatSelectModule } from '@angular/material/select';
-import { MatTableModule } from '@angular/material/table';
 import { RouterLink } from '@angular/router';
-import {
-  createWeekRange,
-  findExerciseDefinition,
-  UnifiedEntryFilterKey,
-} from '@pu-stats/models';
+import { createWeekRange, UnifiedEntryFilterKey } from '@pu-stats/models';
 import { LiveDataStore } from '@pu-stats/data-access';
 import { FilterBarComponent } from '../components/filter-bar/filter-bar.component';
-import { HeatmapComponent } from '../components/heatmap/heatmap.component';
 import { PreviewBannerComponent } from '../components/preview-banner/preview-banner.component';
-import { SetsDistributionComponent } from '../components/sets-distribution/sets-distribution.component';
-import { StatsChartComponent } from '../components/stats-chart/stats-chart.component';
-import { TypePieComponent } from '../components/type-pie/type-pie.component';
 import { AnalysisStore } from '../analysis.store';
-import { exerciseDisplayName } from '../i18n/exercise-display-names';
+import { kindDisplayName } from '../i18n/exercise-display-names';
 import { PageHeaderComponent } from '../../core/page-header/page-header.component';
+import { AnalysisGroupViewComponent } from './analysis-group-view.component';
 
 interface ExerciseKindOption {
   value: UnifiedEntryFilterKey;
@@ -43,21 +33,15 @@ interface ExerciseKindOption {
   providers: [AnalysisStore],
   imports: [
     MatButtonModule,
-    MatButtonToggleModule,
     MatCardModule,
     MatFormFieldModule,
     MatIconModule,
     MatSelectModule,
-    MatTableModule,
     RouterLink,
-    DecimalPipe,
+    AnalysisGroupViewComponent,
     FilterBarComponent,
-    HeatmapComponent,
     PageHeaderComponent,
     PreviewBannerComponent,
-    SetsDistributionComponent,
-    StatsChartComponent,
-    TypePieComponent,
   ],
   template: `
     <main class="page-wrap">
@@ -128,239 +112,7 @@ interface ExerciseKindOption {
         </mat-card>
       }
 
-      <app-stats-chart
-        [series]="store.chartSeries()"
-        [granularity]="store.granularity()"
-        [rangeMode]="store.rangeMode()"
-        [from]="store.from()"
-        [to]="store.to()"
-        [entries]="store.rows()"
-        [dayChartMode]="store.resolvedDayChartMode()"
-        (dayChartModeChange)="store.setDayChartMode($event)"
-      />
-
-      <section class="grid">
-        <mat-card>
-          <mat-card-header>
-            <mat-card-title i18n="@@analysis.bestStreaksTitle"
-              >Bestwerte & Streaks</mat-card-title
-            >
-          </mat-card-header>
-          <mat-card-content class="best-grid">
-            <div>
-              <strong i18n="@@analysis.bestSingleEntry"
-                >Bestwert Einzel-Eintrag:</strong
-              >
-              <div>{{ store.bestSingleEntry()?.reps ?? 0 }} Reps</div>
-            </div>
-            <div>
-              <strong i18n="@@analysis.bestDay">Bester Tag:</strong>
-              <div>
-                {{ store.bestDay()?.date ?? '—' }} ·
-                {{ store.bestDay()?.total ?? 0 }} Reps
-              </div>
-            </div>
-            @if (store.bestSingleSet()) {
-              <div>
-                <strong i18n="@@analysis.bestSingleSet"
-                  >Bestes Einzel-Set:</strong
-                >
-                <div>
-                  {{ store.bestSingleSet() }}
-                  <span i18n="@@analysis.repsUnit">Wdh.</span>
-                </div>
-              </div>
-            }
-            <div>
-              <strong i18n="@@analysis.currentStreak">Aktuelle Streak:</strong>
-              <div>
-                <ng-container i18n="@@analysis.streakDays"
-                  >{{ store.currentStreak() }} Tage</ng-container
-                >
-              </div>
-            </div>
-            <div>
-              <strong i18n="@@analysis.longestStreak">Längste Streak:</strong>
-              <div>
-                <ng-container i18n="@@analysis.streakDays"
-                  >{{ store.longestStreak() }} Tage</ng-container
-                >
-              </div>
-            </div>
-          </mat-card-content>
-        </mat-card>
-
-        <mat-card>
-          <mat-card-header>
-            <mat-card-title i18n="@@analysis.typesTitle"
-              >Typen (Anteile)</mat-card-title
-            >
-          </mat-card-header>
-          <mat-card-content>
-            @defer (hydrate on viewport) {
-              <app-type-pie [data]="typeBreakdownDisplay()" />
-            }
-          </mat-card-content>
-        </mat-card>
-
-        @if (store.avgSetSize()) {
-          <mat-card>
-            <mat-card-header>
-              <mat-card-title i18n="@@analysis.avgSetSizeTitle"
-                >&#x2300; Set-Gr&#x00F6;&#x00DF;e</mat-card-title
-              >
-            </mat-card-header>
-            <mat-card-content class="kpi-big">
-              <b>{{ store.avgSetSize() }}</b>
-              <span i18n="@@analysis.repsPerSet">Reps / Set</span>
-            </mat-card-content>
-          </mat-card>
-        }
-
-        @if (store.setsDistribution().length) {
-          <mat-card>
-            <mat-card-header>
-              <mat-card-title i18n="@@analysis.setsDistTitle"
-                >Sets-Verteilung</mat-card-title
-              >
-            </mat-card-header>
-            <mat-card-content>
-              @defer (hydrate on viewport) {
-                <app-sets-distribution [data]="store.setsDistribution()" />
-              }
-            </mat-card-content>
-          </mat-card>
-        }
-      </section>
-
-      <mat-card class="heatmap-full">
-        <mat-card-header>
-          <mat-card-title i18n="@@analysis.heatmapTitle"
-            >Heatmap (Wochentag/Uhrzeit)</mat-card-title
-          >
-          <mat-button-toggle-group
-            [value]="heatmapMode()"
-            (change)="heatmapMode.set($event.value)"
-            class="heatmap-toggle"
-            aria-label="Heatmap-Modus auswählen"
-            i18n-aria-label="@@analysis.heatmapToggleAriaLabel"
-          >
-            <mat-button-toggle value="reps" i18n="@@analysis.heatmapReps"
-              >Reps</mat-button-toggle
-            >
-            <mat-button-toggle value="sets" i18n="@@analysis.heatmapSets"
-              >Sets</mat-button-toggle
-            >
-          </mat-button-toggle-group>
-        </mat-card-header>
-        <mat-card-content class="heatmap-wrap">
-          @defer (hydrate on viewport) {
-            <app-heatmap [entries]="store.rows()" [mode]="heatmapMode()" />
-          }
-        </mat-card-content>
-      </mat-card>
-
-      <section class="trends-section" data-testid="analysis-trends-section">
-        <mat-card>
-          <mat-card-header>
-            <mat-card-title i18n="@@analysis.weekTrendTitle"
-              >Wochentrend</mat-card-title
-            >
-            <mat-card-subtitle i18n="@@analysis.weekTrendSubtitle"
-              >Letzte 8 Wochen</mat-card-subtitle
-            >
-          </mat-card-header>
-          <mat-card-content>
-            @defer (on viewport) {
-              <mat-table [dataSource]="store.weekTrend()">
-                <ng-container matColumnDef="label">
-                  <mat-header-cell *matHeaderCellDef i18n="@@analysis.weekCol"
-                    >Woche</mat-header-cell
-                  >
-                  <mat-cell *matCellDef="let row">{{ row.label }}</mat-cell>
-                </ng-container>
-                <ng-container matColumnDef="total">
-                  <mat-header-cell *matHeaderCellDef i18n="@@analysis.repsCol"
-                    >Reps</mat-header-cell
-                  >
-                  <mat-cell *matCellDef="let row">{{ row.total }}</mat-cell>
-                </ng-container>
-                <ng-container matColumnDef="avgSetsPerEntry">
-                  <mat-header-cell
-                    *matHeaderCellDef
-                    i18n="@@analysis.avgSetsCol"
-                    >&#x2300; Sets</mat-header-cell
-                  >
-                  <mat-cell *matCellDef="let row">{{
-                    row.avgSetsPerEntry !== null &&
-                    row.avgSetsPerEntry !== undefined
-                      ? (row.avgSetsPerEntry | number)
-                      : '—'
-                  }}</mat-cell>
-                </ng-container>
-                <mat-header-row
-                  *matHeaderRowDef="trendColumnsWithSets"
-                ></mat-header-row>
-                <mat-row
-                  *matRowDef="let row; columns: trendColumnsWithSets"
-                ></mat-row>
-              </mat-table>
-            } @placeholder {
-              <div class="trend-placeholder" aria-hidden="true"></div>
-            }
-          </mat-card-content>
-        </mat-card>
-
-        <mat-card>
-          <mat-card-header>
-            <mat-card-title i18n="@@analysis.monthTrendTitle"
-              >Monatstrend</mat-card-title
-            >
-            <mat-card-subtitle i18n="@@analysis.monthTrendSubtitle"
-              >Letzte 6 Monate</mat-card-subtitle
-            >
-          </mat-card-header>
-          <mat-card-content>
-            @defer (on viewport) {
-              <mat-table [dataSource]="store.monthTrend()">
-                <ng-container matColumnDef="label">
-                  <mat-header-cell *matHeaderCellDef i18n="@@analysis.monthCol"
-                    >Monat</mat-header-cell
-                  >
-                  <mat-cell *matCellDef="let row">{{ row.label }}</mat-cell>
-                </ng-container>
-                <ng-container matColumnDef="total">
-                  <mat-header-cell *matHeaderCellDef i18n="@@analysis.repsCol"
-                    >Reps</mat-header-cell
-                  >
-                  <mat-cell *matCellDef="let row">{{ row.total }}</mat-cell>
-                </ng-container>
-                <ng-container matColumnDef="avgSetsPerEntry">
-                  <mat-header-cell
-                    *matHeaderCellDef
-                    i18n="@@analysis.avgSetsCol"
-                    >&#x2300; Sets</mat-header-cell
-                  >
-                  <mat-cell *matCellDef="let row">{{
-                    row.avgSetsPerEntry !== null &&
-                    row.avgSetsPerEntry !== undefined
-                      ? (row.avgSetsPerEntry | number)
-                      : '—'
-                  }}</mat-cell>
-                </ng-container>
-                <mat-header-row
-                  *matHeaderRowDef="trendColumnsWithSets"
-                ></mat-header-row>
-                <mat-row
-                  *matRowDef="let row; columns: trendColumnsWithSets"
-                ></mat-row>
-              </mat-table>
-            } @placeholder {
-              <div class="trend-placeholder" aria-hidden="true"></div>
-            }
-          </mat-card-content>
-        </mat-card>
-      </section>
+      <app-analysis-group-view />
     </main>
   `,
   styles: `
@@ -370,24 +122,6 @@ interface ExerciseKindOption {
       padding: 16px;
       display: grid;
       gap: 16px;
-    }
-    .grid {
-      display: grid;
-      grid-template-columns: repeat(auto-fit, minmax(320px, 1fr));
-      gap: 16px;
-    }
-    mat-table {
-      width: 100%;
-    }
-
-    mat-card-content {
-      overflow: auto;
-    }
-
-    .best-grid {
-      display: grid;
-      grid-template-columns: repeat(2, minmax(180px, 1fr));
-      gap: 12px;
     }
 
     .filter-section {
@@ -419,62 +153,14 @@ interface ExerciseKindOption {
     }
 
     @media (max-width: 900px) {
-      .filter-section {
-        margin-inline: -12px;
-        padding: 10px 12px;
-      }
-    }
-
-    .chart-card {
-      margin-bottom: 16px;
-    }
-
-    .heatmap-full {
-      width: 100%;
-    }
-    .heatmap-wrap {
-      overflow: auto;
-      width: 100%;
-      min-height: 400px;
-    }
-    /* Reserve roughly the rendered table height so revealing the
-       deferred trend block doesn't shift the rest of the page down. */
-    .trend-placeholder {
-      min-height: 320px;
-    }
-
-    @media (max-width: 900px) {
       .page-wrap {
         padding: 12px;
         gap: 12px;
       }
-      .grid {
-        grid-template-columns: 1fr;
-        gap: 12px;
+      .filter-section {
+        margin-inline: -12px;
+        padding: 10px 12px;
       }
-      .best-grid {
-        grid-template-columns: 1fr 1fr;
-      }
-    }
-
-    .kpi-big {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      gap: 4px;
-      padding: 16px 0;
-    }
-    .kpi-big b {
-      font-size: 2.5rem;
-      line-height: 1;
-    }
-    .kpi-big span {
-      opacity: 0.7;
-      font-size: 0.85rem;
-    }
-
-    .heatmap-toggle {
-      margin-left: auto;
     }
 
     .empty-cta mat-card-content {
@@ -500,20 +186,6 @@ interface ExerciseKindOption {
     }
     .empty-cta a {
       margin-left: auto;
-    }
-
-    @media (max-width: 600px) {
-      .best-grid {
-        grid-template-columns: 1fr;
-      }
-      .heatmap-full mat-card-header {
-        flex-direction: column;
-        align-items: flex-start;
-        gap: 12px;
-      }
-      .heatmap-toggle {
-        margin-left: 0;
-      }
     }
   `,
 })
@@ -553,7 +225,7 @@ export class AnalysisPageComponent {
     }
     return merged.map((value) => ({
       value,
-      label: this.kindLabel(value),
+      label: kindDisplayName(value),
     }));
   });
 
@@ -565,29 +237,6 @@ export class AnalysisPageComponent {
     if (this.store.kinds().length > 0) return true;
     return this.store.kindOptionsRaw().some((k) => k !== 'pushup');
   });
-
-  /**
-   * Resolves the bare-id labels emitted by `store.typeBreakdown()` (in
-   * multi-kind mode) into localised display names. Pushup-only mode
-   * returns the breakdown unchanged because the store already produces
-   * locale-aware variant names there.
-   */
-  readonly typeBreakdownDisplay = computed(() => {
-    const kinds = this.store.kinds();
-    const isKindMode =
-      kinds.length > 0 && !(kinds.length === 1 && kinds[0] === 'pushup');
-    const data = this.store.typeBreakdown();
-    if (!isKindMode) return data;
-    return data.map((d) => ({
-      ...d,
-      label: this.kindLabel(d.id as UnifiedEntryFilterKey),
-    }));
-  });
-
-  readonly trendColumns = ['label', 'total'];
-  readonly trendColumnsWithSets = ['label', 'total', 'avgSetsPerEntry'];
-
-  readonly heatmapMode = signal<'reps' | 'sets'>('reps');
 
   /**
    * Only reveal the empty-state CTA after the entries resource has resolved.
@@ -665,18 +314,5 @@ export class AnalysisPageComponent {
     const qIndex = url.indexOf('?');
     if (qIndex === -1) return '';
     return url.slice(qIndex);
-  }
-
-  private kindLabel(value: UnifiedEntryFilterKey): string {
-    if (value === 'pushup') {
-      return $localize`:@@exercise.category.pushup:Liegestütze`;
-    }
-    const def = findExerciseDefinition(value);
-    if (def) return exerciseDisplayName(def.id);
-    // Legacy id without a catalog entry (e.g. an exercise removed from
-    // the catalog whose entries still live in Firestore) — surface a
-    // generic localised label rather than the raw `category.name` id so
-    // the legend never reads like a developer string.
-    return $localize`:@@analysis.kindUnknown:Andere Übung`;
   }
 }
